Refactored project structure (no content changes)
