+++ components/Base/Navbar/Navbar.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/colors";
+@import "../../../styles/colors";
 
 
 .context {
