+++ components/_plot/Editor/Editor.module.scss
