+++ components/_plot/Editor/Editor.stories.tsx
@@ -1,11 +1,11 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 import Editor from "./Editor";
 import {Provider} from "react-redux";
-import {store} from "../../redux/store";
+import {store} from "../../../redux/store";
 
 
 export default {
-    title: 'components/Editor',
+    title: 'components/_plot/Editor',
     component: Editor
 } as ComponentMeta<typeof Editor>;
